Initial commit
I've already made the pokemon.model. It is an abstract class.

It features healthPoints, attackPoints, enhancement modifiers, a number, name, and whether it can evolve.

There are also pokemon types, and String messages (a.k.a. descriptions and information) on the pokemon.

Finally there are getters and setters for the data members at the beginning of the class.