ABOUT THE INITIAL COMMIT...
The change I've made is only to serve as a way for me to commit my message about the initial commit. I'll fix this error later.

Here is the message:

The initial commit I've just made is NOT entirely correct!
The pokemon.model is NOT a class!
It is a package holding the abstract class that I created.

The abstract class is called "Pokemon".
The rest of the initial commit is correct.