Fixed errors and created methods in PokedexController. Created FaraamKnight as a pokemon. Created an interface called Steel.
PokedexController:
Added parameters in isInt and isDouble methods, and fixed the errors there.
Created the updatePokemon method with parameters being index as an integer and data as a String array.
Created a String array method called buildPokedexText.

FaraamKnight:
Extends the abstract class Pokemon and implements the interface Steel.
FaraamKnight has a number and a name, but no given moves.

Steel:
Nothing but an interface for now.